Decouple OAuthManager behind interface
